+++ decidim-comments/app/events/decidim/comments/comment_created_event.rb
@@ -2,48 +2,25 @@
 
 module Decidim
   module Comments
-    class CommentCreatedEvent < Decidim::Events::BaseEvent
-      include Decidim::Events::EmailEvent
-      include Decidim::Events::NotificationEvent
-
-      def email_subject
-        I18n.t(
-          "decidim.comments.events.comment_created.#{comment_type}.email_subject",
-          resource_title: resource_title,
-          resource_url: resource_locator.url(url_params),
-          author_name: author.name
-        )
-      end
-
-      def email_intro
-        I18n.t(
-          "decidim.comments.events.comment_created.#{comment_type}.email_intro",
-          resource_title: resource_title
-        ).html_safe
-      end
-
-      def email_outro
-        I18n.t(
-          "decidim.comments.events.comment_created.#{comment_type}.email_outro",
-          resource_title: resource_title
-        )
-      end
-
-      def notification_title
-        I18n.t(
-          "decidim.comments.events.comment_created.#{comment_type}.notification_title",
-          resource_title: resource_title,
-          resource_path: resource_locator.path(url_params),
-          author_nickname: author.nickname,
-          author_name: author.name,
-          author_path: author.profile_path
-        ).html_safe
-      end
+    class CommentCreatedEvent < Decidim::Events::SimpleEvent
+      include Decidim::Events::AuthorEvent
 
       private
 
+      def i18n_scope
+        "decidim.events.comments.comment_created.#{comment_type}"
+      end
+
+      def resource_path
+        resource_locator.path(url_params)
+      end
+
+      def resource_url
+        resource_locator.url(url_params)
+      end
+
       def author
-        @author ||= Decidim::UserPresenter.new(comment.author)
+        comment.author
       end
 
       def comment